Added upload image feature
