+++ bheta-dashboard/src/app/(pwa)/pwa/utils/postImages.ts
@@ -1,0 +1,10 @@
+export const createObjectURL = (file: File): string => {
+    return URL.createObjectURL(file);
+  };
+  
+  export const validateImageFile = (file: File | null): string | null => {
+    if (!file) {
+      return 'Please select an image to upload.';
+    }
+    return null;
+  };
