+++ bheta-dashboard/src/app/page.tsx
@@ -1,10 +1,13 @@
-import ImageUpload from "./(pwa)/pwa/Uploadimage/page";
+import SignUpForm from "./(dashboard)/dashboard/sign-up/page";
+import ImageUpload from "./(pwa)/pwa/component/uploadimage/page";
+
 
 
 export default function Home() {
   return (
     <div>
-      <ImageUpload />
+     <ImageUpload />
+      {/* <SignUpForm /> */}
     </div>
   );
 }
